requirements.txt
numpy
xgboost
Flask
notebook
dvc
gunicorn
sklearn
pandas
PyYAML
evidently
imblearn
dill
matplotlib
tqdm
seaborn
python-box==6.0.2
ensure==1.0.2
joblib
types-PyYAML
scipy
-e .